Moved more functions into `common`
