+++ lib/admin/route/admin.instance.create.property/admin.instance.create.property.controller.js
@@ -10,12 +10,16 @@ const {
 } = require('@ministryofjustice/fb-runner-node/lib/page/set-errors/set-errors')
 
 const {
-  getServiceSchema,
   temporaryFile,
   expandInstance
 } = require('~/fb-editor-node/service-data/service-data')
 
 const instancePropertyController = require('~/fb-editor-node/admin/route/admin.instance.property/admin.instance.property.controller')
+
+const {
+  getSchemaProperties,
+  transformValueToDataType
+} = require('~/fb-editor-node/admin/common')
 
 function populateError (e, pageInstance) {
   const errors = []
@@ -73,23 +77,13 @@ async function postValidation (pageInstance, pageData, postRedirect) {
   // console.log('lib/admin/route/admin.instance.create.property.controller.postValidation()')
 
   const _id = pageData.getUserDataProperty('_id')
+  const {
+    _type
+  } = pageData.getUserDataProperty('sourceInstance')
+
+  const parentProperty = pageData.getUserDataProperty('parentProperty')
   const property = pageData.getUserDataProperty('property')
-  let value = pageData.getUserDataProperty('value')
-
-  const sourceInstance = pageData.getUserDataProperty('sourceInstance')
-
-  // TODO: this block replicates code in admin.instance.property
-  // and it should be handled when the values are received
-  const schema = getServiceSchema(sourceInstance._type)
-
-  const propertyType = schema.properties[property].type
-  if (propertyType === 'number') {
-    value *= 1
-  } else if (propertyType === 'boolean') {
-    value = value === 'true'
-  } else if (propertyType === 'array' || propertyType === 'object' || value === undefined || value.match(/^\{[\s\S]*\}$/)) {
-    value = JSON.parse(value)
-  }
+  const value = transformValueToDataType(pageData.getUserDataProperty('value'), getSchemaProperties(_type, parentProperty), property)
 
   const file = pageData.getUserDataProperty('temporaryFile')
 
@@ -101,7 +95,7 @@ async function postValidation (pageInstance, pageData, postRedirect) {
   /*
    *  Compare as strings
    */
-  if (JSON.stringify(value) === JSON.stringify(previousValue)) return postRedirect.success()
+  if (JSON.stringify(value) === JSON.stringify(previousValue)) return postRedirect && postRedirect.success()
 
   /*
    *  Save the current value
