+++ lib/admin/route/admin.instance/admin.instance.controller.js
@@ -1021,7 +1021,7 @@ async function postValidation (pageInstance, pageData, postRedirect) {
       /*
        *  Compare as strings
        */
-      if (instance === sourceInstance) return postRedirect.success()
+      if (instance === sourceInstance) return postRedirect && postRedirect.success()
     }
 
     /*
